More refactoring, move fetchEvents to event history instead of the page layout
